+++ forge/src/main/java/net/haizor/fancydyes/forge/oculus/DyeRenderType.java
@@ -1,0 +1,5 @@
+package net.haizor.fancydyes.forge.oculus;
+
+public interface DyeRenderType {
+    boolean isDye();
+}