+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-* Fix compatibility issue with Geckolib
+* Fix compat issue with Oculus on forge